feat: support alpaca fantom & multichain refactor
